Split NBGL layout implementation regarding Keyboard & Keypad
