+++ lib_nbgl/include/nbgl_layout.h
@@ -26,8 +26,7 @@ extern "C" {
 /*********************
  *      DEFINES
  *********************/
-#define NO_MORE_OBJ_ERROR -3
-#define NBGL_NO_TUNE      NB_TUNES
+#define NBGL_NO_TUNE NB_TUNES
 
 #ifdef HAVE_SE_TOUCH
 #define NB_MAX_SUGGESTION_BUTTONS 4
